add user define function
