+++ lib/renderer/function.js
@@ -147,12 +147,11 @@ module.exports = function (Renderer) {
                 }
                 else {
                     var me = this;
-                    var params = '{';
+                    var params = [];
                     node.attrs.forEach(function (nod) {
-                        params += nod.key.value + ': ' + (nod.value ? me._getExpr(nod.value) : '') + ',';
+                        params.push(nod.key.value + ': ' + (nod.value ? me._getExpr(nod.value) : ''));
                     });
-                    params = params.slice(0, params.length - 1) + '}'
-                    return '__func["__fn__' + funcName + '"](' + params +')' + (flag == 2 ? '' : ';');
+                    return '__func["__fn__' + funcName + '"]({' + params.join(',') + '})' + (flag == 2 ? '' : ';');
                 }
             }
             else {
