+++ lib/renderer/index.js
@@ -14,7 +14,7 @@ function addPhpFuncs(engine, jsTpl) {
     
     var arr = [];
     var obj = {};
-    var matches = jsTpl.match(/__f\.(.*?)\(/g);
+    var matches = jsTpl.match(/__f\..+\(/g);
     if (matches) {
         matches.forEach(
             function (item) {
@@ -23,10 +23,31 @@ function addPhpFuncs(engine, jsTpl) {
         );
         utils.excludeItem(arr).forEach(
             function (p) {
-                obj[p] = engine.func[p];
+                obj[p] = engine.phpfunc[p];
             }
         );
         return 'var __f=' + utils.toFuncString(obj) + ';';
+    }
+    return '';
+}
+
+function addUserDefinedFuncs(engine, jsTpl) {
+    var arr = [];
+    var obj = {};
+    var matches = jsTpl.match(/"__fn__.+"/g);
+    console.log(matches)
+    if (matches) {
+        matches.forEach(
+            function (item) {
+                arr.push(item.slice(7, item.length - 1));
+            }
+        );
+        utils.excludeItem(arr).forEach(
+            function (p) {
+                obj[p] = engine.func[p];
+            }
+        );
+        return 'var __func=' + utils.toFuncString(obj) + ';';
     }
     return '';
 }
@@ -94,7 +115,7 @@ Renderer.prototype = {
                 +     '}'
                 + '};'
                 + me._init(ast) + '\nif(__tn){__h="";__func["__fn__"+__tn](__da); return __h;}' + '\nreturn __h;'
-            jsTpl = addClosure(addPhpFuncs(this.engine, jsTpl) + jsTpl);
+            jsTpl = addClosure(/*addUserDefinedFuncs(this.engine, jsTpl) + */addPhpFuncs(this.engine, jsTpl) + jsTpl);
         }
         else {
             jsTpl += me._init(ast, flag);
