+++ lib/renderer/literal.js
@@ -175,7 +175,7 @@ module.exports = function (Renderer) {
             var pstr = '';
             var me = this;
             var ps = node.params;
-            ps.forEach(function (p) {
+            ps && ps.forEach(function (p) {
                 pstr += me._getExpr(p) + ((p != ps[ps.length - 1]) ? ',' : '');
             });
             return '__f.' + node.name + '(' + pstr + ')'
