+++ lib/utils.js
@@ -76,14 +76,13 @@ module.exports = {
     },
 
     toFuncString: function (obj) {
-        var str = '{';
+        var ret = [];
         for (var p in obj) {
-            if (obj.hasOwnProperty(p)) {
-                str += p + ':' + obj[p].toString() + ',';
+            if (obj.hasOwnProperty(p) && obj[p]) {
+                ret.push(p + ':' + obj[p].toString());
             }
         }
-        str = str.substr(0, str.length - 1);
-        return str += '}'
+        return '{' + ret.join(',') + '}'
     },
 
     excludeItem: function (arr) {
